Client/requirements.txt
fastapi
uvicorn
gradio_client
gunicorn
biopython
